refactor: 自定义模板资源工作空间隔离

+++ backend/src/main/java/io/metersphere/service/IssueTemplateService.java
@@ -1,5 +1,7 @@
 package io.metersphere.service;
 
+import io.metersphere.base.domain.CustomField;
+import io.metersphere.base.domain.CustomFieldTemplate;
 import io.metersphere.base.domain.IssueTemplate;
 import io.metersphere.base.domain.IssueTemplateExample;
 import io.metersphere.base.mapper.IssueTemplateMapper;
@@ -11,6 +13,7 @@ import io.metersphere.commons.utils.ServiceUtils;
 import io.metersphere.controller.request.BaseQueryRequest;
 import io.metersphere.controller.request.UpdateIssueTemplateRequest;
 import io.metersphere.i18n.Translator;
+import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -32,6 +35,9 @@ public class IssueTemplateService {
     @Resource
     CustomFieldTemplateService customFieldTemplateService;
 
+    @Resource
+    CustomFieldService customFieldService;
+
     public void add(UpdateIssueTemplateRequest request) {
         checkExist(request);
         IssueTemplate template = new IssueTemplate();
@@ -39,6 +45,10 @@ public class IssueTemplateService {
         template.setId(UUID.randomUUID().toString());
         template.setCreateTime(System.currentTimeMillis());
         template.setUpdateTime(System.currentTimeMillis());
+        if (request.getSystem() == null) {
+            request.setSystem(false);
+        }
+        template.setGlobal(false);
         issueTemplateMapper.insert(template);
         customFieldTemplateService.create(request.getCustomFields(), template.getId(),
                 TemplateConstants.FieldTemplateScene.ISSUE.name());
@@ -51,24 +61,77 @@ public class IssueTemplateService {
 
     public void delete(String id) {
         issueTemplateMapper.deleteByPrimaryKey(id);
+        customFieldTemplateService.deleteByTemplateId(id);
     }
 
     public void update(UpdateIssueTemplateRequest request) {
-        checkExist(request);
-        customFieldTemplateService.deleteByTemplateId(request.getId());
-        IssueTemplate template = new IssueTemplate();
-        BeanUtils.copyBean(template, request);
-        template.setUpdateTime(System.currentTimeMillis());
-        issueTemplateMapper.updateByPrimaryKeySelective(template);
-        customFieldTemplateService.create(request.getCustomFields(), request.getId(),
-                TemplateConstants.FieldTemplateScene.ISSUE.name());
+        if (request.getGlobal() != null && request.getGlobal()) {
+            // 如果是全局字段，则创建对应工作空间字段
+            add(request);
+        } else {
+            checkExist(request);
+            customFieldTemplateService.deleteByTemplateId(request.getId());
+            IssueTemplate template = new IssueTemplate();
+            BeanUtils.copyBean(template, request);
+            template.setUpdateTime(System.currentTimeMillis());
+            issueTemplateMapper.updateByPrimaryKeySelective(template);
+            customFieldTemplateService.create(request.getCustomFields(), request.getId(),
+                    TemplateConstants.FieldTemplateScene.ISSUE.name());
+        }
+    }
+
+    /**
+     * 获取该工作空间的系统模板
+     * - 如果没有，则创建该工作空间模板，并关联默认的字段
+     * - 如果有，则更新原来关联的 fieldId
+     * @param customField
+     */
+    public void handleSystemFieldCreate(CustomField customField) {
+        IssueTemplate workspaceSystemTemplate = getWorkspaceSystemTemplate(customField.getWorkspaceId());
+        if (workspaceSystemTemplate == null) {
+            createTemplateWithUpdateField(customField.getWorkspaceId(), customField);
+        } else {
+            updateRelateWithUpdateField(workspaceSystemTemplate, customField);
+        }
+    }
+
+    private IssueTemplate getWorkspaceSystemTemplate(String workspaceId) {
+        IssueTemplateExample example = new IssueTemplateExample();
+        example.createCriteria()
+                .andWorkspaceIdEqualTo(workspaceId)
+                .andSystemEqualTo(true);
+        List<IssueTemplate> issueTemplates = issueTemplateMapper.selectByExampleWithBLOBs(example);
+        if (CollectionUtils.isNotEmpty(issueTemplates)) {
+            return issueTemplates.get(0);
+        }
+        return null;
+    }
+
+    private void createTemplateWithUpdateField(String workspaceId, CustomField customField) {
+        UpdateIssueTemplateRequest request = new UpdateIssueTemplateRequest();
+        IssueTemplate issueTemplate = new IssueTemplate();
+        issueTemplate.setName("default");
+        issueTemplate.setPlatform(TemplateConstants.IssueTemplatePlatform.metersphere.name());
+        issueTemplate.setGlobal(false);
+        issueTemplate.setWorkspaceId(workspaceId);
+        BeanUtils.copyBean(request, issueTemplate);
+        List<CustomFieldTemplate> systemFieldCreateTemplate =
+                customFieldTemplateService.getSystemFieldCreateTemplate(customField, TemplateConstants.FieldTemplateScene.ISSUE.name());
+        request.setCustomFields(systemFieldCreateTemplate);
+        add(request);
+    }
+
+    private void updateRelateWithUpdateField(IssueTemplate template, CustomField customField) {
+        CustomField globalField = customFieldService.getGlobalFieldByName(customField.getName());
+        customFieldTemplateService.updateFieldIdByTemplate(template.getId(), globalField.getId(), customField.getId());
     }
 
     private void checkExist(IssueTemplate issueTemplate) {
         if (issueTemplate.getName() != null) {
             IssueTemplateExample example = new IssueTemplateExample();
             IssueTemplateExample.Criteria criteria = example.createCriteria();
-            criteria.andNameEqualTo(issueTemplate.getName());
+            criteria.andNameEqualTo(issueTemplate.getName())
+                    .andWorkspaceIdEqualTo(issueTemplate.getWorkspaceId());
             if (StringUtils.isNotBlank(issueTemplate.getId())) {
                 criteria.andIdNotEqualTo(issueTemplate.getId());
             }
